requirements.txt
PySide6
pyperclip
Pillow
pywin32